Move User related things to its very own package in order to do not mix them with the Video related ones (we don't want to go all in for the Bounded Context and Modules/sub-domains option, but at least we want to have some separation of concerns)
