+++ src/main/tv/codely/scala_http_api/Routes.scala
@@ -3,8 +3,8 @@ package tv.codely.scala_http_api
 import akka.http.scaladsl.model.{ContentTypes, HttpEntity}
 import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server.Route
-import tv.codely.scala_http_api.domain.User
-import tv.codely.scala_http_api.infrastructure.UserMarshaller._
+import tv.codely.scala_http_api.user.domain.User
+import tv.codely.scala_http_api.user.infrastructure.UserMarshaller._
 
 object Routes {
   private val systemUsers = Seq(
