--- src/main/tv/codely/scala_http_api/domain/UserName.scala
@@ -1,3 +1,0 @@
-package tv.codely.scala_http_api.domain
-
-case class UserName(value: String)
