+++ src/main/tv/codely/scala_http_api/user/domain/User.scala
@@ -1,4 +1,4 @@
-package tv.codely.scala_http_api.domain
+package tv.codely.scala_http_api.user.domain
 
 object User {
   def apply(id: String, name: String): User = User(UserId(id), UserName(name))
