+++ src/test/tv/codely/scala_http_api/shared/infrastructure/stub/StringStub.scala
@@ -1,4 +1,4 @@
-package tv.codely.scala_http_api.infrastructure.stubs
+package tv.codely.scala_http_api.shared.infrastructure.stub
 
 import scala.util.Random
